requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
coverage==3.7.1
hiredis==0.1.6
nose==1.3.4
PyYAML==3.11
redis==2.10.3
tornado==4.1